requirements.txt
streamlit
transformers==4.11.3
python-dotenv
tokenizers==0.10.1
pip==24.0
openai